fix(windows + reorganizing): Made the test pass for windows and reorganized test to have all audio record tests together.
